requirements.txt
Package                Version
---------------------- -------
mysql-connector-python 8.0.25 
pip                    19.2.3 
protobuf               4.21.11
python-dateutil        2.8.2  
pytz                   2022.6 
setuptools             41.2.0 
six                    1.16.0 
